# src/main/resources/openapi/prices-api.yaml
openapi: 3.0.3
info:
  title: Prices API
  description: Returns the applicable price for a product, brand and date.
  version: 1.0.0

paths:
  /api/prices:
    get:
      tags:
        - Prices
      summary: Get applicable price
      operationId: getPrice
      parameters:
        - name: applicationDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
          example: "2020-06-14T10:00:00"
        - name: productId
          in: query
          required: true
          schema:
            type: integer
            format: int64
          example: 35455
        - name: brandId
          in: query
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: Applicable price found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceResponse'
        "400":
          description: Missing or invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceException'
        "404":
          description: No price found for the given inputs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceException'

components:
  schemas:
    PriceResponse:
      type: object
      required:
        - productId
        - brandId
        - priceList
        - startDate
        - endDate
        - price
        - currency
      properties:
        productId:
          type: integer
          format: int64
          example: 35455
        brandId:
          type: integer
          format: int64
          example: 1
        priceList:
          type: integer
          example: 1
        startDate:
          type: string
          format: date-time
          example: "2020-06-14T00:00:00"
        endDate:
          type: string
          format: date-time
          example: "2020-12-31T23:59:59"
        price:
          type: number
          example: 35.50
        currency:
          type: string
          example: "EUR"

    PriceException:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string